+++ archey/test/test_archey_colors.py
@@ -1,17 +1,32 @@
 """Test module for `archey.colors`"""
 
 import unittest
-from unittest.mock import Mock, patch
+from unittest.mock import patch
 
 from archey.colors import ANSI_ECMA_REGEXP, Colors
 
 
-@patch(
-    'archey.colors.Environment',
-    Mock(NO_COLOR=False)  # By default, colors won't be disabled.
-)
 class TestColors(unittest.TestCase):
     """Test cases for the `Colors` (enumeration / utility) class"""
+    def setUp(self):
+        # Skip `Colors.should_color_output` patching for its own testing method.
+        if self.id().endswith('should_color_output'):
+            return
+
+        # By default, colors won't be disabled.
+        self._should_color_output_patch = patch(
+            'archey.colors.Colors.should_color_output',
+            return_value=True
+        )
+        self._should_color_output_patch.start()
+
+    def tearDown(self):
+        # Skip `Colors.should_color_output` patching for its own testing method.
+        if self.id().endswith('should_color_output'):
+            return
+
+        self._should_color_output_patch.stop()
+
     def test_constant_values(self):
         """Test enumeration member instantiation from value"""
         self.assertEqual(Colors((1, 31)), Colors.RED_BRIGHT)
@@ -21,6 +36,44 @@ class TestColors(unittest.TestCase):
         """Simple test case for `__str__` implementation"""
         self.assertEqual(str(Colors.CLEAR), '\x1b[0m')
         self.assertEqual(str(Colors.CYAN_BRIGHT), '\x1b[1;36m')
+
+    def test_should_color_output(self):
+        """Test for `Colors.should_color_output`"""
+        # Clear cache filled by `functools.lru_cache` decorator.
+        Colors.should_color_output.cache_clear()
+
+        with patch('archey.colors.Environment.CLICOLOR_FORCE', True):
+            self.assertTrue(Colors.should_color_output())
+
+        Colors.should_color_output.cache_clear()
+
+        with patch('archey.colors.Environment.NO_COLOR', True):
+            self.assertFalse(Colors.should_color_output())
+
+        Colors.should_color_output.cache_clear()
+
+        with patch('archey.colors.sys.stdout.isatty', return_value=False):
+            with patch('archey.colors.Environment.CLICOLOR', True):
+                self.assertFalse(Colors.should_color_output())
+
+            Colors.should_color_output.cache_clear()
+
+            with patch('archey.colors.Environment.CLICOLOR', False):
+                self.assertFalse(Colors.should_color_output())
+
+            Colors.should_color_output.cache_clear()
+
+        with patch('archey.colors.sys.stdout.isatty', return_value=True):
+            # Default case : STDOUT is a TTY and `CLICOLOR` is (by default) set.
+            with patch('archey.colors.Environment.CLICOLOR', True):
+                self.assertTrue(Colors.should_color_output())
+
+            Colors.should_color_output.cache_clear()
+
+            with patch('archey.colors.Environment.CLICOLOR', False):
+                self.assertFalse(Colors.should_color_output())
+
+            Colors.should_color_output.cache_clear()
 
     def test_escape_code_from_attrs(self):
         """Test for `Colors.escape_code_from_attrs`"""
@@ -68,9 +121,9 @@ class TestColors(unittest.TestCase):
             '\x1b[0nTEST\xde\xad\xbe\xaf'
         )
 
-    def test_no_color(self):
-        """Test `Colors` behavior when `NO_COLOR` is set"""
-        with patch('archey.colors.Environment', Mock(NO_COLOR=True)):
+    def test_color_disabling(self):
+        """Test `Colors` internal behavior when coloration is disabled"""
+        with patch('archey.colors.Colors.should_color_output', return_value=False):
             self.assertFalse(str(Colors.CYAN_NORMAL))